davinciai-backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.3.0

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1

# Redis
redis==5.0.1
aioredis==2.0.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Payments
stripe==7.10.0

# HTTP Client
httpx==0.26.0
aiohttp==3.9.1

# WebSocket
python-socketio==5.10.0
websockets==12.0

# Utilities
python-dotenv==1.0.0
loguru==0.7.2

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
black==23.12.1
flake8==7.0.0
mypy==1.8.0
